requirements.txt
# Core ML libraries
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
lightgbm==4.0.0
xgboost==2.0.3
joblib==1.3.2

# Notebooks
jupyter==1.0.0

# API
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.7

# MLFlow
mlflow==2.9.2

# Visualization (for notebooks)
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.14.1

# Development
pytest==7.4.0
requests==2.31.0
